1_data_pipeline/requirements.txt
openmeteo-requests
requests-cache retry-requests numpy pandas
requests pandas lxml
python-dotenv
holidays